requirements.txt
openai
together
google-generativeai
google-cloud-aiplatform
python-dotenv
